Code V1.0 of DL-ICEEMDAN-PE-GT(SVD) for IEEEM2VIP/requirements.txt
torchviz~=0.0.2
thop~=0.1.1-2209072238
torchinfo~=1.8.0
fvcore~=0.1.5.post20221221
numpy~=1.25.2
scikit-learn~=1.3.0
librosa~=0.10.1
scipy~=1.11.1
pyentrp~=0.9.0
pandas~=2.1.0
hiddenlayer~=0.3
tqdm~=4.65.0
seaborn~=0.12.2
matplotlib~=3.7.2